discordbot/moviebot/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
Deprecated==1.2.13
docopt==0.6.2
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
Js2Py==0.71
multidict==6.0.2
nextcord==2.0.0
opentelemetry-api==1.11.1
opentelemetry-instrumentation==0.32b0
opentelemetry-instrumentation-sqlalchemy==0.32b0
opentelemetry-semantic-conventions==0.32b0
packaging==21.3
pycodestyle==2.8.0
pycparser==2.21
pyjsparser==2.7.1
PyNaCl==1.4.0
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==0.20.0
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
toml==0.10.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
wrapt==1.14.1
yarl==1.7.2
